django_domain_utils/requirements.txt
-r reqs/prod.txt
pylibmc==1.4.0
django-pylibmc==0.5.0